ath79: Add support for OpenMesh OM2P v1
Device specifications:
======================

* Qualcomm/Atheros AR7240 rev 2
* 350/350/175 MHz (CPU/DDR/AHB)
* 32 MB of RAM
* 16 MB of SPI NOR flash
  - 2x 7 MB available; but one of the 7 MB regions is the recovery image
* 2x 10/100 Mbps Ethernet
* 1T1R 2.4 GHz Wi-Fi
* 6x GPIO-LEDs (3x wifi, 2x ethernet, 1x power)
* 1x GPIO-button (reset)
* external h/w watchdog (enabled by default)
* TTL pins are on board (arrow points to VCC, then follows: GND, TX, RX)
* 2x fast ethernet
  - eth0
    + 18-24V passive POE (mode B)
    + used as WAN interface
  - eth1
    + builtin switch port 4
    + used as LAN interface
* 12-24V 1A DC
* external antenna

Flashing instructions:
======================

The u-boot environment must be changed slightly to fix the bootup with the
ath79. The NOR flash device was previously using the identifier ar7240-nor0
on the OM2P v1 on ar71xx. But under ath79, the identifier is now spi0.0 and
cannot be overwritten using a device tree option. ath79 is therefore no
longer able to extract the correct partition informations from the mtdparts
boot arguments.

In the u-boot shell, simply enter:

  setenv set_bootargs_1 'setenv bootargs_1 console=ttyS0,115200 root=31:04 rootfstype=squashfs,jffs2 init=/etc/preinit mtdparts=spi0.0:256k(u-boot),256k(u-boot-env),1280k(custom),${kernel_size_1}k(kernel),${rootfs_size_1}k(rootfs),7168k(inactive),256k(ART) board=OM2P'
  setenv set_bootargs_2 'setenv bootargs_2 console=ttyS0,115200 root=31:05 rootfstype=squashfs,jffs2 init=/etc/preinit mtdparts=spi0.0:256k(u-boot),256k(u-boot-env),1280k(custom),7168k(inactive),${kernel_size_2}k(kernel),${rootfs_size_2}k(rootfs),256k(ART) board=OM2P'
  saveenv
  reset

Various methods can then be used to install the actual image on the flash.
Two easy ones are:

ap51-flash
----------

The tool ap51-flash (https://github.com/ap51-flash/ap51-flash) should be
used to transfer the image to the u-boot when the device boots up.

initramfs from TFTP
-------------------

The serial console must be used to access the u-boot shell during bootup.
It can then be used to first boot up the initramfs image from a TFTP server
(here with the IP 192.168.1.21):

   setenv serverip 192.168.1.21
   setenv ipaddr 192.168.1.1
   tftpboot 0c00000 <filename-of-initramfs-kernel>.bin && bootm $fileaddr

The actual sysupgrade image can then be transferred (on the LAN port) to the
device via

  scp <filename-of-squashfs-sysupgrade>.bin root@192.168.1.1:/tmp/

On the device, the sysupgrade must then be started using

  sysupgrade -n /tmp/<filename-of-squashfs-sysupgrade>.bin

Signed-off-by: Sven Eckelmann <sven@narfation.org>