ath79: Optimize openmesh sysupgrade rootfs copy
The rootfs read (and piping to mtd) in the sysupgrade script is still done
with dd and a block size of 1 byte. This is necessary because the normal
squashfs image is

 (256KB or 64KB) * blocks + 4 bytes

large. This is never divisible by something like 1024 bytes (as used for
the kernel read).

But it is still possible to copy in 1024 byte large chunks when relaxing
the requirement "copy exactly the number of bytes for the rootfs to flash"
to "copy at least the number of bytes for the rootfs to flash". Then it
will definitely include the "End of filesystem marker". But as downside,
also some extra bogus data from the sysupgrade metadata is also copied to
the flash.

This reduces the kernel+rootfs write time on an OM2P-HS v3 on for a 5.4MB
large image from 90s to 30s. The rootfs was 3407876 bytes and the kernel
2097152 bytes large.

Reported-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Signed-off-by: Sven Eckelmann <sven@narfation.org>