ipq40xx: utilize nvmem-cells for openmesh,a62
The calibration data and mac addresses on this device are stored in the
0:ART partition. It is therefore possible to move the code to handle them
directly to the devicetree instead of the various scripts.

But the actual relevant information about the partition layout is provided
by the bootloader via bootargs (mtdparts) and not via the devicetree
itself. The template in the devicetree is only used when booting the
initramfs or to identify the partitions for the nvmem-cells referenced by
the ethernet + wifi devices. The kernel will take care of mapping the
content of the template to the mtdparts partition by its label (and other
properties).

Reported-by: Robert Marko <robert.marko@sartura.hr>
Signed-off-by: Sven Eckelmann <sven@narfation.org>